+++ packages/atomic/src/components/common/facets/timeframe-facet-common.tsx
@@ -326,7 +326,7 @@ export class TimeframeFacetCommon {
         isCollapsed={isCollapsed}
         headingLevel={this.props.headingLevel}
         onToggleCollapse={onToggleCollapse}
-        headerRef={headerFocus.setTarget}
+        headerRef={(el) => headerFocus.setTarget(el)}
       ></FacetHeader>
     );
   }