+++ packages/atomic/src/components/insight/result-lists/atomic-insight-result-list/atomic-insight-result-list.tsx
@@ -11,10 +11,7 @@ import {
   InsightResult,
   buildInsightInteractiveResult,
 } from '../..';
-import {
-  FocusTarget,
-  FocusTargetController,
-} from '../../../../utils/accessibility-utils';
+import {FocusTargetController} from '../../../../utils/accessibility-utils';
 import {
   BindStateToController,
   InitializableComponent,
@@ -62,7 +59,7 @@ export class AtomicInsightResultList
   @State() private resultTemplateRegistered = false;
   @State() public error!: Error;
 
-  @FocusTarget() nextNewResultTarget!: FocusTargetController;
+  private nextNewResultTarget?: FocusTargetController;
 
   /**
    * The spacing of various elements in the result list, including the gap between results, the gap between parts of a result, and the font sizes of different parts in a result.
@@ -119,7 +116,7 @@ export class AtomicInsightResultList
       getResultDisplay: () => this.display,
       getLayoutDisplay: () => this.display,
       getImageSize: () => this.imageSize,
-      nextNewResultTarget: this.nextNewResultTarget,
+      nextNewResultTarget: this.focusTarget,
       loadingFlag: this.loadingFlag,
       getResultListState: () => this.resultListState,
       getResultRenderingFunction: () => this.resultRenderingFunction,
@@ -131,6 +128,13 @@ export class AtomicInsightResultList
           options: {result},
         }),
     });
+  }
+
+  private get focusTarget(): FocusTargetController {
+    if (!this.nextNewResultTarget) {
+      this.nextNewResultTarget = new FocusTargetController(this);
+    }
+    return this.nextNewResultTarget;
   }
 
   public render() {