+++ packages/atomic/src/components/search/atomic-breadbox/atomic-breadbox.tsx
@@ -9,10 +9,7 @@ import {
 } from '@coveo/headless';
 import {Component, h, State, Element, VNode} from '@stencil/core';
 import CloseIcon from '../../../images/close.svg';
-import {
-  FocusTarget,
-  FocusTargetController,
-} from '../../../utils/accessibility-utils';
+import {FocusTargetController} from '../../../utils/accessibility-utils';
 import {getFieldValueCaption} from '../../../utils/field-utils';
 import {
   InitializableComponent,
@@ -78,14 +75,11 @@ export class AtomicBreadbox implements InitializableComponent {
   @State() public error!: Error;
   @State() private isCollapsed = true;
 
-  @FocusTarget()
-  private breadcrumbRemovedFocus!: FocusTargetController;
+  private breadcrumbRemovedFocus?: FocusTargetController;
 
-  @FocusTarget()
-  private breadcrumbShowMoreFocus!: FocusTargetController;
+  private breadcrumbShowMoreFocus?: FocusTargetController;
 
-  @FocusTarget()
-  private breadcrumbShowLessFocus!: FocusTargetController;
+  private breadcrumbShowLessFocus?: FocusTargetController;
 
   public initialize() {
     this.breadcrumbManager = buildBreadcrumbManager(this.bindings.engine);
@@ -99,6 +93,23 @@ export class AtomicBreadbox implements InitializableComponent {
 
   public disconnectedCallback() {
     this.resizeObserver?.disconnect();
+  }
+
+  private get focusTargets() {
+    if (!this.breadcrumbRemovedFocus) {
+      this.breadcrumbRemovedFocus = new FocusTargetController(this);
+    }
+    if (!this.breadcrumbShowLessFocus) {
+      this.breadcrumbShowLessFocus = new FocusTargetController(this);
+    }
+    if (!this.breadcrumbShowMoreFocus) {
+      this.breadcrumbShowMoreFocus = new FocusTargetController(this);
+    }
+    return {
+      breadcrumbRemovedFocus: this.breadcrumbRemovedFocus,
+      breadcrumbShowLessFocus: this.breadcrumbShowLessFocus,
+      breadcrumbShowMoreFocus: this.breadcrumbShowMoreFocus,
+    };
   }
 
   private get breadcrumbs() {
@@ -213,7 +224,7 @@ export class AtomicBreadbox implements InitializableComponent {
             if (isLastBreadcrumb) {
               this.bindings.store.state.resultList?.focusOnFirstResultAfterNextSearch();
             } else if (this.numberOfBreadcrumbs > 1) {
-              this.breadcrumbRemovedFocus.focusAfterSearch();
+              this.focusTargets.breadcrumbRemovedFocus.focusAfterSearch();
             }
 
             this.lastRemovedBreadcrumbIndex = index;
@@ -221,10 +232,10 @@ export class AtomicBreadbox implements InitializableComponent {
           }}
           ref={(ref) => {
             if (this.lastRemovedBreadcrumbIndex === index) {
-              this.breadcrumbRemovedFocus.setTarget(ref);
+              this.focusTargets.breadcrumbRemovedFocus.setTarget(ref);
             }
             if (this.firstExpandedBreadcrumbIndex === index) {
-              this.breadcrumbShowMoreFocus.setTarget(ref);
+              this.focusTargets.breadcrumbShowMoreFocus.setTarget(ref);
             }
           }}
         >
@@ -279,7 +290,7 @@ export class AtomicBreadbox implements InitializableComponent {
       <li key="show-more">
         <Button
           ref={(ref) => {
-            this.breadcrumbShowLessFocus.setTarget(ref!);
+            this.focusTargets.breadcrumbShowLessFocus.setTarget(ref!);
             this.showMore = ref!;
           }}
           part="show-more"
@@ -288,7 +299,7 @@ export class AtomicBreadbox implements InitializableComponent {
           onClick={() => {
             this.firstExpandedBreadcrumbIndex =
               this.numberOfBreadcrumbs - this.numberOfCollapsedBreadcrumbs;
-            this.breadcrumbShowMoreFocus.focusOnNextTarget();
+            this.focusTargets.breadcrumbShowMoreFocus.focusOnNextTarget();
             this.isCollapsed = false;
           }}
         ></Button>
@@ -306,7 +317,7 @@ export class AtomicBreadbox implements InitializableComponent {
           text={this.bindings.i18n.t('show-less')}
           class="p-2 btn-pill"
           onClick={() => {
-            this.breadcrumbShowLessFocus.focusOnNextTarget();
+            this.focusTargets.breadcrumbShowLessFocus.focusOnNextTarget();
             this.isCollapsed = true;
           }}
         ></Button>
@@ -330,7 +341,9 @@ export class AtomicBreadbox implements InitializableComponent {
             this.bindings.store.state.resultList?.focusOnFirstResultAfterNextSearch();
           }}
           ref={
-            isFocusTarget ? this.breadcrumbRemovedFocus.setTarget : undefined
+            isFocusTarget
+              ? this.focusTargets.breadcrumbRemovedFocus.setTarget
+              : undefined
           }
         ></Button>
       </li>