+++ packages/atomic/src/components/search/facets/atomic-color-facet/atomic-color-facet.tsx
@@ -14,7 +14,6 @@ import {
 import {Component, h, State, Prop, VNode, Element} from '@stencil/core';
 import {
   AriaLiveRegion,
-  FocusTarget,
   FocusTargetController,
 } from '../../../../utils/accessibility-utils';
 import {
@@ -234,14 +233,11 @@ export class AtomicColorFacet
   @Prop({mutable: true})
   public customSort: string[] | string = '[]';
 
-  @FocusTarget()
-  private showLessFocus!: FocusTargetController;
+  private showLessFocus?: FocusTargetController;
 
-  @FocusTarget()
-  private showMoreFocus!: FocusTargetController;
+  private showMoreFocus?: FocusTargetController;
 
-  @FocusTarget()
-  private headerFocus!: FocusTargetController;
+  private headerFocus?: FocusTargetController;
 
   @AriaLiveRegion('facet-search')
   protected facetSearchAriaMessage!: string;
@@ -273,6 +269,24 @@ export class AtomicColorFacet
       numberOfActiveValues: () => this.numberOfSelectedValues,
     });
     this.initializeDependenciesManager();
+  }
+
+  private get focusTargets() {
+    if (!this.showLessFocus) {
+      this.showLessFocus = new FocusTargetController(this);
+    }
+    if (!this.showMoreFocus) {
+      this.showMoreFocus = new FocusTargetController(this);
+    }
+    if (!this.headerFocus) {
+      this.headerFocus = new FocusTargetController(this);
+    }
+
+    return {
+      showLessFocus: this.showLessFocus,
+      showMoreFocus: this.showMoreFocus,
+      headerFocus: this.headerFocus,
+    };
   }
 
   public disconnectedCallback() {
@@ -318,14 +332,14 @@ export class AtomicColorFacet
         i18n={this.bindings.i18n}
         label={this.label}
         onClearFilters={() => {
-          this.headerFocus.focusAfterSearch();
+          this.focusTargets.headerFocus.focusAfterSearch();
           this.facet.deselectAll();
         }}
         numberOfActiveValues={this.numberOfSelectedValues}
         isCollapsed={this.isCollapsed}
         headingLevel={this.headingLevel}
         onToggleCollapse={() => (this.isCollapsed = !this.isCollapsed)}
-        headerRef={this.headerFocus.setTarget}
+        headerRef={(el) => this.focusTargets.headerFocus.setTarget(el)}
       ></FacetHeader>
     );
   }
@@ -382,8 +396,10 @@ export class AtomicColorFacet
             onClick={onClick}
             searchQuery={this.facetState.facetSearch.query}
             buttonRef={(element) => {
-              isShowLessFocusTarget && this.showLessFocus.setTarget(element);
-              isShowMoreFocusTarget && this.showMoreFocus.setTarget(element);
+              isShowLessFocusTarget &&
+                this.focusTargets.showLessFocus.setTarget(element);
+              isShowMoreFocusTarget &&
+                this.focusTargets.showMoreFocus.setTarget(element);
             }}
           >
             <FacetValueLabelHighlight
@@ -403,8 +419,10 @@ export class AtomicColorFacet
             onClick={onClick}
             searchQuery={this.facetState.facetSearch.query}
             buttonRef={(element) => {
-              isShowLessFocusTarget && this.showLessFocus.setTarget(element);
-              isShowMoreFocusTarget && this.showMoreFocus.setTarget(element);
+              isShowLessFocusTarget &&
+                this.focusTargets.showLessFocus.setTarget(element);
+              isShowMoreFocusTarget &&
+                this.focusTargets.showMoreFocus.setTarget(element);
             }}
           >
             <div
@@ -491,11 +509,11 @@ export class AtomicColorFacet
         i18n={this.bindings.i18n}
         onShowMore={() => {
           this.resultIndexToFocusOnShowMore = this.facet.state.values.length;
-          this.showMoreFocus.focusAfterSearch();
+          this.focusTargets.showMoreFocus.focusAfterSearch();
           this.facet.showMoreValues();
         }}
         onShowLess={() => {
-          this.showLessFocus.focusAfterSearch();
+          this.focusTargets.showLessFocus.focusAfterSearch();
           this.facet.showLessValues();
         }}
         canShowLessValues={this.facetState.canShowLessValues}